+++ src/app/_components/_templates/simple-table-manual/helpers.ts
