+++ src/app/_components/_templates/simple-table-manual/simple-table-manual.component.html
@@ -1,0 +1,12 @@
+<ng-container  *ngIf="bVisible$">
+    <div class="simple-table-wrapper">
+        <table mat-table [dataSource]="rowData" class="mat-elevation-z8">
+            <ng-container *ngFor="let dCol of processedFields" [matColumnDef]="dCol.config.name">
+                <th mat-header-cell *matHeaderCellDef> {{ dCol.config.label }}</th>
+                <td mat-cell *matCellDef="let element"> {{ element[dCol.config.name] }} </td>
+            </ng-container>
+            <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+        </table>
+    </div>
+</ng-container>
