+++ src/app/_components/_templates/simple-table-manual/simple-table-manual.component.scss
@@ -1,0 +1,155 @@
+@import "../../../_shared/styles.scss";
+
+/* Adapted from list-view.component.scss */
+
+  .simple-table-wrapper {
+    width: 100%;
+    margin-top: 0.5rem;
+    margin-bottom: 0.5rem;
+  }
+  
+  table {
+    width: 100%;
+  }
+
+  ::ng-deep .mat-sort-header-content {
+    white-space: normal;
+  }
+
+  ::ng-deep td.mat-cell {
+    white-space: normal;
+
+  }
+
+  ::ng-deep th.mat-header-cell, td.mat-cell, td.mat-footer-cell {
+    padding: 0.188rem 0.188rem !important;
+  }
+
+  ::ng-deep .mat-button {
+    padding: 0px;
+    text-align: left;
+  }
+
+  .mat-icon-button {
+    width: fit-content;
+  }
+
+  .mat-row .mat-cell {
+      text-align: left;
+  }
+
+  .psdk-mat-header {
+    white-space: normal;
+    padding-right: 5px;
+  }
+
+  .psdk-mat-header-button {
+    display: inline-grid;
+  }
+
+  .psdk-mat-header-arrow {
+    display: inline-grid;
+    vertical-align: middle;
+  }
+
+  .psdk-mat-header-filter {
+    display: inline-grid;
+    vertical-align: middle;
+  }
+
+  .psdk-full-width {
+    width: 100%;
+  }   
+
+  .psdk-search {
+    padding-left: 0.625rem;
+  }
+
+  .psdk-icon-search {
+      vertical-align: sub;
+      padding: 0rem 0.125rem;
+      min-width: unset;
+      width: 1.1rem;
+  }
+
+  .psdk-outer-div-in-form {
+    display: inline-grid;
+  }
+
+  .psdk-inner-div-in-form {
+    height: auto;
+    position: relative;
+    width: 100%;
+    overflow: auto;
+    max-height: 550px;
+    min-height: auto;
+  }
+
+  .psdk-list-view-svg-icon {
+    width: 1.4rem;
+    display: inline-block;
+    vertical-align: middle;
+    filter: $app-primary-color-filter;
+  }
+
+  .psdk-filter-svg-icon {
+    width: 1.0rem;
+    display: inline-block;
+    vertical-align: middle;
+    filter: $app-neutral-color-filter;
+  }
+
+  .psdk-arrow-svg-icon {
+    width: 1.0rem;
+    display: inline-block;
+    vertical-align: middle;
+    filter: $app-neutral-color-filter;
+  }
+
+  .psdk-filter-popover {
+    display: table;
+    margin: auto;
+    min-width: 100px;
+    background-color: white;
+    border: 1px solid black;
+    border-radius: 10px;
+    padding: 20px;
+    box-shadow: 0 0 10px 3px #777;
+    position: absolute;
+    z-index: 99;
+}
+
+.psdk-dialog-background {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 100%;
+  background-color: rgba(100,100,100, 0.4);
+  position: fixed;
+  z-index: 999;
+  top: 0px;
+  left: 0px;
+}
+
+.psdk-modal-file-top {
+  display: table;
+  margin: auto;
+  min-width: 150px;
+  background-color: white;
+  border: 1px solid black;
+  border-radius: 10px;
+  padding: 20px;
+  box-shadow: 0 0 10px 3px #777;
+}
+
+  tr.mat-row {
+    cursor: pointer;
+  }
+
+.psdk-data-readonly {
+    margin-top: 0.625rem;
+    width: 100%;
+  }
+
