+++ src/app/_components/_templates/simple-table/simple-table.component.html
@@ -3,18 +3,7 @@
     <app-field-group-template [configProps$]="fieldGroupProps" [pConn$]="pConn$" [formGroup$]="formGroup$"></app-field-group-template>
   </div>
   <ng-template #table>
-    <div class="simple-table-wrapper">
-      <table mat-table [dataSource]="rowData" class="mat-elevation-z8">
-    
-        <ng-container *ngFor="let dCol of processedFields" [matColumnDef]="dCol.config.name">
-          <th mat-header-cell *matHeaderCellDef> {{ dCol.config.label }}</th>
-          <td mat-cell *matCellDef="let element"> {{ element[dCol.config.name] }} </td>
-        </ng-container>
-      
-        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-      </table>
-    </div>
+    <app-simple-table-manual [pConn$]="pConn$" [formGroup$]="formGroup$"></app-simple-table-manual>
   </ng-template>
 </ng-container>
