+++ src/app/app.module.ts
@@ -141,6 +141,7 @@ import { FieldGroupTemplateComponent } from './_components/_templates/field-grou
 import { FieldGroupListComponent } from './_components/_templates/field-group-list/field-group-list.component';
 import { FieldGroupUtils } from './_helpers/field-group-utils';
 import { MdInputBlurEventDirective } from './_directives/md-input-blur-event.directive';
+import { SimpleTableManualComponent } from './_components/_templates/simple-table-manual/simple-table-manual.component';
 @NgModule({
   declarations: [
     AppComponent,
@@ -238,7 +239,8 @@ import { MdInputBlurEventDirective } from './_directives/md-input-blur-event.dir
     MultiReferenceReadonlyComponent,
     FieldGroupTemplateComponent,
     FieldGroupListComponent,
-    MdInputBlurEventDirective
+    MdInputBlurEventDirective,
+    SimpleTableManualComponent
   ],
   imports: [
     BrowserModule,
